Cambios generales y de base de datos
Se separa más la logica del back en distintos archivos y se cambia para que se use email para autenticar, además de añadir el tipo de usuario en el payload
