+++ login.js
@@ -1,15 +1,14 @@
 const express = require("express");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-const dotenv = require("dotenv");
+const db = require('./database');
 
 
-dotenv.config();
 const login = express();
 login.use(express.json());
 login.post('/login', async (req, res) => {
-    const { DNI, Contrasenya } = req.body;
-    db.query('SELECT * FROM usuaris WHERE DNI = ?', [DNI], async (error, results) => {
+    const { Email, Contrasenya } = req.body;
+    db.query('SELECT * FROM usuaris WHERE Email = ?', [Email], async (error, results) => {
         if (error) {
             console.log("Error:", error);
             return res.status(500).json({ error: 'Error en el servidor' });
@@ -22,7 +21,7 @@ login.post('/login', async (req, res) => {
         if (!contrasenyaValida) {
             return res.status(400).json({ error: 'Credenciales incorrectas' });
         }
-        const token = jwt.sign({ id: usuario.id_usuari }, process.env.JWT_SECRET, { expiresIn: '1h' });
+        const token = jwt.sign({ id: usuario.id_usuari, tipus: usuario.TipusUsuari }, process.env.JWT_SECRET, { expiresIn: '1h' });
         res.json({ token });
     });
 });
